# specs/001-ai-monitoring-center/contracts/api.openapi.yaml
openapi: 3.1.0
info:
  title: Интеллектуальный центр PR/GR-мониторинга — внутренний API
  version: 0.1.0
  description: >
    Контракт между backend и frontend. Фиксируется ДО начала кодирования и служит
    границей между дорожками двух инженеров. Изменение контракта — только через
    правку spec.md и этого файла, никогда в одностороннем порядке.
    Авторизация в MVP отсутствует по решению заказчика.
servers:
  - url: http://localhost:8000/api
...
paths:
...
  /items/{id}:
    get:
      tags: [items]
      summary: Карточка материала с разложением оценки и заземлением (FR-024, FR-025)
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ItemDetail" } } } }
        "404": { description: Not found }
    patch:
      tags: [items]
      summary: Правка полей материала (FR-040, FR-044)
      description: Изменённые поля фиксируются в журнале правок и не перезаписываются переобработкой (FR-042, FR-043).
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:      { type: string }
                summary:    { type: string }
                topic:      { type: string, enum: [regulatory, reputation, competitors, trends] }
                tags:       { type: array, items: { type: string } }
                user_note:  { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ItemDetail" } } } }
  /items/{id}/assessment:
    patch:
      tags: [items]
      summary: Правка отдельных баллов; индекс и категория пересчитываются немедленно (FR-041)
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scores]
              properties:
                scores:
                  type: object
                  additionalProperties: { type: integer, minimum: 0, maximum: 3 }
                escalation_flags: { type: array, items: { type: string } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Assessment" } } } }
        "422": { description: Балл вне диапазона 0–3 или неизвестный критерий }
  /items/{id}/hide:
    post:
      tags: [items]
      summary: Скрыть или вернуть публикацию (FR-045)
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                hidden: { type: boolean }
                reason: { type: string, description: "Обязательно при hidden=true" }
      responses: { "204": { description: OK } }
  /items:
    post:
      tags: [items]
      summary: Ручное добавление материала (FR-009)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, raw_text]
              properties:
                title:        { type: string }
                raw_text:     { type: string }
                url:          { type: string }
                published_at: { type: string, format: date-time }
      responses: { "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/ItemDetail" } } } } }
...
  /sources:
    get:
      tags: [sources]
      summary: Список источников со статусом сбора
      responses: { "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Source" } } } } } }
    post:
      tags: [sources]
      summary: Добавить источник (FR-003)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, url]
              properties:
                type:     { type: string, enum: [rss, web, telegram] }
                url:      { type: string }
                title:    { type: string }
                category: { type: string, enum: [media, regulator, telegram] }
                item_type: { type: string, enum: [news, act], description: "Тип материалов; по умолчанию выводится из категории источника" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Source" } } } }
        "422": { description: URL недоступен или тип не соответствует содержимому }
  /sources/{id}:
    patch:
      tags: [sources]
      summary: Включить, отключить или переименовать источник (FR-004)
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:             { type: string }
                is_active:         { type: boolean }
                poll_interval_min: { type: integer }
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Source" } } } } }
    delete:
      tags: [sources]
      summary: Удалить источник; собранные материалы сохраняются (FR-004)
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { "204": { description: OK } }
  /sources/{id}/collect:
    post:
      tags: [sources]
      summary: Немедленный запуск сбора — используется на демонстрации
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { "202": { description: Accepted } }
...
  /stories/{id}:
    get:
      tags: [items]
      summary: Карточка события со всеми источниками и ссылками на оригиналы (FR-060)
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Story" } } } } }
  /stories/{id}/split:
    post:
      tags: [items]
      summary: Разъединить ошибочно объединённый кластер (FR-062)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: reason, in: query, required: false, schema: { type: string } }
      responses: { "204": { description: OK } }
...
components:
  schemas:
    ItemCard:
      type: object
      required: [id, title, summary, item_type, index_value, final_category, url]
      properties:
        id:              { type: integer }
        title:           { type: string }
        summary:         { type: string, description: "3–5 предложений (FR-010)" }
        entities:        { $ref: "#/components/schemas/Entities" }
        item_type:       { type: string, enum: [news, act] }
        topic:           { type: string }
        index_value:     { type: number, description: "ИВ или ИА, вычислено кодом (FR-016)" }
        final_category:  { type: string }
        escalation_flags: { type: array, items: { type: string } }
        url:             { type: string, description: "ссылка на оригинал (FR-023)" }
        source:          { type: object, properties: { id: { type: integer }, title: { type: string } } }
        published_at:    { type: string, format: date-time }
        is_partial_text: { type: boolean }
        is_edited:       { type: boolean, description: "есть пользовательские правки" }
        story:           { type: object, nullable: true, properties: { id: { type: integer }, item_count: { type: integer } } }
        act_id:          { type: integer, nullable: true }
        tags:            { type: array, items: { type: string } }
    ItemDetail:
      allOf:
        - $ref: "#/components/schemas/ItemCard"
        - type: object
          properties:
            raw_text:   { type: string }
            claims:     { type: array, items: { $ref: "#/components/schemas/Claim" } }
            assessment: { $ref: "#/components/schemas/Assessment" }
            user_note:  { type: string, nullable: true }
            revisions:  { type: array, items: { $ref: "#/components/schemas/Revision" } }
            machine_summary: { type: string, nullable: true, description: "Последняя машинная версия для сравнения с правкой" }
            machine_assessment: { allOf: [{ $ref: "#/components/schemas/Assessment" }], nullable: true }
    Claim:
      type: object
      description: Заземление утверждения саммари на оригинал (FR-012, FR-025)
      required: [statement, quote]
      properties:
        statement:  { type: string }
        quote:      { type: string }
        char_start: { type: integer }
        char_end:   { type: integer }
    Entities:
      type: object
      description: Ключевые сущности (FR-011)
      properties:
        who:          { type: array, items: { type: string } }
        what:         { type: string }
        when:         { type: string }
        consequences: { type: string }
    Assessment:
      type: object
      required: [scheme, scores, rationales, index_value, category, final_category, author]
      properties:
        scheme:      { type: string, enum: [npa_k1_k6, news_h1_h4] }
        scores:      { type: object, additionalProperties: { type: integer, minimum: 0, maximum: 3 } }
        rationales:  { type: object, additionalProperties: { type: string } }
        index_value: { type: number }
        category:    { type: string }
        escalation_flags: { type: array, items: { type: string } }
        final_category:   { type: string }
        author:      { type: string, enum: [ai, human] }
        model:       { type: string, nullable: true }
        prompt_version: { type: string, nullable: true }
        created_at:  { type: string, format: date-time }
...
    Source:
      type: object
      required: [id, type, url, item_type, is_active]
      properties:
        id:                { type: integer }
        type:              { type: string, enum: [rss, web, telegram, manual] }
        url:               { type: string }
        title:             { type: string }
        category:          { type: string }
        item_type:         { type: string, enum: [news, act] }
        is_active:         { type: boolean }
        poll_interval_min: { type: integer }
        last_polled_at:    { type: string, format: date-time, nullable: true }
        last_error:        { type: string, nullable: true }
        item_count:        { type: integer }
...
    Revision:
      type: object
      properties:
        field:      { type: string }
        old_value:  {}
        new_value:  {}
        author:     { type: string }
        reason:     { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    Story:
      type: object
      required: [id, canonical_title, items]
      properties:
        id: { type: integer }
        canonical_title: { type: string }
        fact_summary: { type: string }
        first_seen_at: { type: string, format: date-time }
        item_count: { type: integer }
        was_split_by_user: { type: boolean }
        items:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              title: { type: string }
              url: { type: string }
              source: { type: object }
              published_at: { type: string, format: date-time }